+++ client/src/components/layout/Sidebar.vue
@@ -109,7 +109,7 @@
     </nav>
 
     <div class="p-4 border-t border-gray-800 text-sm text-gray-400">
-      <div>v0.37.8c-alpha</div>
+      <div>v0.37.8d-alpha</div>
     </div>
   </aside>
 </template>